src/requirements.txt
requests==2.32.3
fastapi==0.110.0
uvicorn==0.27.1
python-jose==3.3.0
boto3==1.34.50
mysql-connector-python==8.3.0
